changed the name of all the js files to end with .js  instead of .mjs to fix bug with jest and unit-testing
