+++ src/js/api/auth/index.js
@@ -1,0 +1,2 @@
+export * from "./login.js";
+export * from "./register.js";
