--- src/js/api/auth/index.mjs
@@ -1,2 +1,0 @@
-export * from "./login.mjs";
-export * from "./register.mjs";
