+++ src/js/api/requests/get.js
@@ -1,16 +1,16 @@
-import { callApiWith } from "../apiCall.mjs";
-import { API_LISTINGS, API_PROFILES } from "../../constants/index.mjs";
-import { PARAM_bids } from "../../constants/params.mjs";
-import { userFeedback } from "../../ui/components/errors/userFeedback.mjs";
+import { callApiWith } from "../apiCall.js";
+import { API_LISTINGS, API_PROFILES } from "../../constants/index.js";
+import { PARAM_bids } from "../../constants/params.js";
+import { userFeedback } from "../../ui/components/errors/userFeedback.js";
 
-let page = 1;
+// let page = 1;
 
-const errorMessage = "We are having some trouble with our servers, please wait and try again later";
-const loadMoreBtn = document.getElementById("loadMore");
-const loader = document.getElementById("loader");
+// const errorMessage = "We are having some trouble with our servers, please wait and try again later";
+// const loadMoreBtn = document.getElementById("loadMore");
+// const loader = document.getElementById("loader");
 
 export async function getData(url) {
-  const errorContainer = document.getElementById("userFeedback");
+  // const errorContainer = document.getElementById("userFeedback");
 
   const response = await callApiWith(url);
   if (response.ok) {
