+++ src/js/bootstrapImport.js
@@ -1,3 +1,3 @@
 // Import the Bootstrap bundle
 // This includes Popper and all of Bootstrap's JS plugins.
-import "../../node_modules/bootstrap/dist/js/bootstrap.bundle.min.js";
+import "bootstrap/dist/js/bootstrap.bundle.min.js";
