+++ src/js/constants/api_secret.js
@@ -1,0 +1,1 @@
+export const API_KEY = "d2abe438-b85e-4837-bb83-2678d612f606";
