--- src/js/constants/constants.mjs
@@ -1,2 +1,0 @@
-export const searchInput = document.querySelector("#searchbar");
-export const newPostInput = document.querySelector(".new-post-input");
