--- src/js/constants/index.mjs
@@ -1,3 +1,0 @@
-export * from "./api.mjs";
-export * from "./api_secret.mjs";
-export * from "./constants.mjs";
