+++ src/js/storage/index.js
@@ -1,0 +1,2 @@
+export * from "./load.js";
+export * from "./save.js";
