--- src/js/storage/index.mjs
@@ -1,2 +1,0 @@
-export * from "./load.mjs";
-export * from "./save.mjs";
